+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Require Redis 2.6. (ntalbott)
 * Add deprecation support. (ntalbott)
 * Add proper API for accessing indexes by name. (ntalbott)
+* Add lookup of multi index values. (ntalbott)
 
 ### 1.5.0